% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applymetrics.R
\name{applymetrics}
\alias{applymetrics}
\title{applymetrics}
\usage{
applymetrics(data, n = 4, sf, lb = 0.2, hb = 15, epochsize = 5)
}
\arguments{
\item{data}{Three column acceleration time series}

\item{n}{Order of frequency filter to be applied}

\item{sf}{Sample frequency (Hertz)}

\item{lb}{Lower boundary (Hertz) of frequency filter}

\item{hb}{Higher boundary (Hertz) of frequency filter}

\item{epochsize}{epochsize to be used}
}
\value{
Data.frame with acceleration metrics values at epoch level
}
\description{
'applymetrics' adaptation of the g.applymetrics function in GGIR
}
