% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_domain_functions.R
\name{computeCrossCorrelations}
\alias{computeCrossCorrelations}
\title{computeCrossCorrelations}
\usage{
computeCrossCorrelations(data, plot)
}
\arguments{
\item{data}{An object of class "spec"}
}
\value{
List object consisting of two elements: \item{correlationMatrix}{A correlation matrix representing the maximum cross-correlations between the signals} \item{lags}{A matrix that presents the lags corresponding to the lag in which the maximum cross-correlation occurred}
}
\description{
'computeCrossCorrelations' computes the maximum cross-correlations between two signals and the corresponding lag
}
