% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_domain_functions.R
\name{createBoxplot}
\alias{createBoxplot}
\title{createBoxplot}
\usage{
createBoxplot(
  data,
  freqBins,
  outcome = c("domFreq", "meanPSD"),
  group = c("brand", "dynamic_range"),
  sampling_rate = c("low", "high", ""),
  orientation_analyses = FALSE
)
}
\arguments{
\item{data}{A data set containing the comparison values for the experiment}

\item{freqBins}{A vector that indicates the cut-offs for the frequency bins}

\item{outcome}{String containing the type of the comparison value. One of c("domFreq", "meanPSD")}

\item{group}{String to indicate the grouping variable. One of c("brand", "dynamic_range")}

\item{orientation_analyses}{Boolean to indicate if the analysis was for the ms_bag experiment, default = FALSE}
}
\value{
Boxplot between or within devices for the outcome (y-axis) per frequency bin (x-axis)
}
\description{
'createBoxplot' creates a boxplot between or within devices for the dominant frequency or the mean power spectral density per frequency bin
}
