% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_domain_functions.R
\name{dataframeShapeComparisonBeween}
\alias{dataframeShapeComparisonBeween}
\title{dataframeShapeComparisonBeween}
\usage{
dataframeShapeComparisonBeween(crossCorrelations, specifications, between)
}
\arguments{
\item{crossCorrelations}{A list consisting of two elements: \item{correlationMatrix}{A correlation matrix representing the maximum cross-correlations between the signals} \item{lags}{A matrix that presents the lags corresponding to the lag in which the maximum cross-correlation occurred}}

\item{specifications}{A data.frame object that represents specifications of the signal, including serial_number, brand, and dynamic range}

\item{between}{A string one of c("brands", dynamic_ranges")}
}
\value{
A data.frame with the following variables: \item{correlation}{Representing the cross-correlations between the two signals} \item{lag}{Representing the lag corresonding to the maximum correlation between the two signals} \item{between1}{Representing the brand/dynamic range of the first signal} \item{between2}{Representing the brand/dynamic range of the second signal} \item{device1}{Representing the serial number of the device that recorded the first signal} \item{device2}{Representing the serial number of the device that recorded the second signal} \item{same_brand}{An indicator of correlations that can be selected for the within (1) or between(0) brand/dynamic range comparison}
}
\description{
'dataframeShapeComparisonBeween' derives a data.frame object required for between and within brand comparison of the cross-correlations
}
