% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_domain_functions.R
\name{deriveComparisonValues}
\alias{deriveComparisonValues}
\title{deriveComparisonValues}
\usage{
deriveComparisonValues(spectrum, freqBins)
}
\arguments{
\item{spectrum}{An object of class "spec"}

\item{freqBins}{Vector containing frequency values that define the frequency bins (can be derived using the function 'deriveBins')}
}
\value{
Data.frame consisting of two columns: \item{domFreq}{The dominant frequency of the signal in each frequency bin, i.e. frequency associated with the maximum value} \item{meanPSD}{The mean power spectral density of the signal in each frequency bin}
}
\description{
'deriveComparisonValues' derives the dominant frequency and the mean power spectral density from a frequency spectrum
}
