% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_domain_functions.R
\name{deriveMeanSpectrum}
\alias{deriveMeanSpectrum}
\title{deriveMeanSpectrum}
\usage{
deriveMeanSpectrum(spectra)
}
\arguments{
\item{spectra}{A list containing the raw spectra over which the mean will be derived}
}
\value{
A list containing two following elements:
\item{x}{Vector of frequencies at which the spectral density (y) is estimated}
\item{y}{Vector of estimates of the spectral density at frequencies corresponding to frequencies (x)}
}
\description{
'deriveMeanSpectrum' derives the mean frequency spectrum from the derived spectra
}
