% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_domain_functions.R
\name{deriveSpectrum}
\alias{deriveSpectrum}
\title{deriveSpectrum}
\usage{
deriveSpectrum(
  x,
  sampling_rate,
  file_name,
  plot = TRUE,
  datadir,
  XLIM = c(0, 5)
)
}
\arguments{
\item{x}{Vector that contains the values of the signal (i.e., acceleration measured in the shaking direction)}

\item{sampling_rate}{Integer that indicates the sampling rate of the signal (i.e., the average number of samples obtained in one second in Hz)}

\item{file_name}{String to indicate the name of the plot (i.e., "spectrum" result in the following name for plot 1 "spectrum_1.jpeg").}

\item{plot}{Boolean, default is TRUE the derived spectrum will be plotted}

\item{datadir}{String that defines the path to save the plot}

\item{XLIM}{Vector that limits the frequency content of the plot, default is c(0, 5) to limit the frequency content to 5 Hz (as 250 rpm / 60 = 4.1667 Hz is the expected max)}
}
\value{
An object of class "spec", which is a list containing at least the following components:
\item{freq}{Vector of frequencies at which the spectral density is estimated.}
\item{spec}{Vector (for univariate series) or matrix (for multivariate series) of estimates of the spectral density at frequencies corresponding to freq.}
\item{coh}{NULL for univariate series. For multivariate time series, a matrix containing the squared coherency between different series.}
\item{phase}{NULL for univariate series. For multivariate time series a matrix containing the cross-spectrum phase between different series.}
\item{series}{The name of the time series.}
\item{snames}{For multivariate input, the names of the component series.}
\item{method}{The method used to calculate the spectrum.}
}
\description{
'deriveSpectrum' derives the frequency spectrum of a signal
}
