% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{loaddata}
\alias{loaddata}
\title{loaddata}
\usage{
loaddata(
  path,
  brand,
  experiment,
  windows = TRUE,
  experimentfile,
  actigraph_preprocessing = TRUE
)
}
\arguments{
\item{path}{Path to the root of the experimental data (rawdatadir)}

\item{brand}{Sensor brand: "Actigraph", "Activpal", "Acttrust", "Axivity", "GENEActiv", or "MOX".}

\item{experiment}{Experiment to load: "timer_check", "ms_hfcr", "ms_lfcr", "ms_hfmr", "ms_lfmr", or "box".}

\item{windows}{Boolean, if TRUE then windows from the data_description file will be selected, FALSE: complete data file will be loaded;}

\item{experimentfile}{xlsx file with protocol description}

\item{actigraph_preprocessing}{Boolean,if  TRUE then sleepmode segments that jump by >1 sec are filled up using the function fill_sleep}
}
\value{
A list of: \item{data}{List of data.frames with the accelerometer time series where each list item represents 1 recording} \item{specifications}{Specifications for each recording}
}
\description{
'loaddata' Loads data
}
