% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_domain_functions.R
\name{pairwisePlots}
\alias{pairwisePlots}
\title{pairwisePlots}
\usage{
pairwisePlots(correlationMatrix, data, threshold)
}
\arguments{
\item{correlationMatrix}{A matrix that indicates for which pairs plots need to be derived}

\item{data}{A list that includes the timeseries data}

\item{threshold}{A double number indicating the threshold cross-correlation for which pairwise plots will be generated}
}
\value{
Plot including two time series the timeseries of the device first in the title is presented in black
}
\description{
'pairwisePlots' generates a plot including two timeseries
}
