% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.activpal.R
\name{read.activpal}
\alias{read.activpal}
\title{read.activpal}
\usage{
read.activpal(file.name.and.path)
}
\arguments{
\item{file.name.and.path}{Full file path to Activpal file (.csv format)}
}
\value{
Data.frame with acceleration timeseries and timestamps
}
\description{
'read.activpal' function adapted from package (activalProcessing), as the original function (activpal.file.reader) is not suitable for the raw data
}
