% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_data.R
\name{structure_data}
\alias{structure_data}
\title{structure_data}
\usage{
structure_data(
  brands,
  experiments,
  rawdatadir,
  outputdir,
  experimentfile = c()
)
}
\arguments{
\item{brands}{Character vector with sensor brands to consider, which may include: "Actigraph", "Activpal", "Acttrust", "Axivity", "GENEActiv", and "MOX".}

\item{experiments}{Character vector of experiments to load, which may include: "timer_check", "ms_hfcr", "ms_lfcr", "ms_hfmr", "ms_lfmr", and "box".}

\item{rawdatadir}{Path to the root of the experimental data (rawdatadir)}

\item{outputdir}{Output directory}

\item{experimentfile}{xlsx file with protocol description, defaults to file stored inside the code}
}
\value{
No output is given, data is stored in .RData files
}
\description{
'structure_data' is the first step in the processing pipeline and loads the raw unstructured data and structures it to ease post-procressing
}
