% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocalibration.R
\name{autocalibration}
\alias{autocalibration}
\title{autocalibration}
\usage{
autocalibration(data, sf, printsummary = TRUE, brand)
}
\arguments{
\item{data}{data.frame with accelerometer time series}

\item{sf}{Sample frequency in Hertz}

\item{printsummary}{Boolean to indicate whether calibration summary should be printed}

\item{brand}{Character of sensor brand: "Actigraph", "Activpal", "Acttrust", "Axivity", "GENEActiv", or "MOX".}
}
\value{
List of objects identical to g.calibrate function R package GGIR
}
\description{
'autocalibration' simplified version of auto-calibration from R package GGIR without temperature and without loading the data in blocks
}
