% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getspecs.R
\name{getspecs}
\alias{getspecs}
\title{getspecs}
\usage{
getspecs(brand, data, experimentfile = c(), experiment)
}
\arguments{
\item{brand}{Sensor brand: "ActiGraph", "activPAL", "Acttrust", "Axivity", "GENEActiv", or "MOX".}

\item{data}{Data object}

\item{experimentfile}{.xlsx file with protocol description, defaults to file stored inside the code}

\item{experiment}{Experiment to load: "timer_check", "ms_hrcr", "ms_lrcr", "ms_mrcr", ms_hrmr", "ms_lrmr", "ms_bag" or "box".}
}
\value{
specifications, a data.frame with serial_number, sampling_rate, and dynamic_range
}
\description{
'getspecs' Called from within loaddata to extract recording specifications
}
