% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwindows.R
\name{getwindows}
\alias{getwindows}
\title{getwindows}
\usage{
getwindows(brand, experiment, path, data, experimentfile = c())
}
\arguments{
\item{brand}{Sensor brand: "ActiGraph", "activPAL", "Acttrust", "Axivity", "GENEActiv", or "MOX".}

\item{experiment}{Experiment to load: "timer_check", "ms_hrcr", "ms_lrcr", "ms_mrcr", "ms_hrmr", "ms_lrmr", ms_bag or "box".}

\item{path}{Path to the root of the experimental data (rawdatadir)}

\item{data}{Data object}

\item{experimentfile}{.xlsx file with protocol description, defaults to file stored inside the code}
}
\value{
List of data.frames with the accelerometer time series where each list item represents 1 recording
}
\description{
'getwindows' Called from within loaddata to extract specific time windows
}
